feat: prepare files, move inputs to shared, add update button
