+++ client/src/App.tsx
@@ -6,6 +6,7 @@ import { Titlepage } from './Pages/Titlepage'
 import { RegisterForm } from './Pages/Register'
 import { LoginForm } from './Pages/Login'
 import { WakeUp } from './Pages/WakeUpPage'
+import { Updatepage } from './Pages/Updatepage'
 
 function App() {
   return (
@@ -16,6 +17,7 @@ function App() {
         <Route path="/recipes" element={<Homepage />} />
         <Route path="/recipes/:id" element={<Recipepage />} />
         <Route path="/add-recipe" element={<Createpage />} />
+        <Route path="/update-recipe/:id" element={<Updatepage />} />
         <Route path="/register" element={<RegisterForm />} />
         <Route path="/login" element={<LoginForm />} />
       </Routes>
