+++ client/src/Pages/Updatepage.tsx
@@ -1,0 +1,12 @@
+import { Box } from '@chakra-ui/react'
+import { Footer } from '../Components/Shared/Footer/Footer'
+import { Header } from '../Components/Shared/Header/Header'
+
+export function Updatepage() {
+  return (
+    <Box bg="#d0ffd5">
+      <Header></Header>
+      <Footer></Footer>
+    </Box>
+  )
+}
